Bugfix/open bugfixes

+++ src/django/VLE/models.py
@@ -457,7 +457,7 @@ class Entry(models.Model):
     createdate = models.DateTimeField(
         default=now,
     )
-    grade = models.IntegerField(
+    grade = models.FloatField(
         default=None,
         null=True,
     )
